[ACC-259] UI clean + Update filename distrib

+++ ariane_relsrv/server/restful.py
@@ -769,11 +769,12 @@ class RestCheckout(Resource):
                         dpath = dpath[0] + '/'
                         os.chdir(os.path.join(project_path, dpath))
                         flag_dir = False
-                    print("PRINT ", os.getcwd(), "git checkout " + df.path[len(dpath):]+df.name)
-                    if df.is_versioned():
+                    if df.is_versioned() and "master.SNAPSHOT" not in df.name:
                         if os.path.isfile(df.path[len(dpath):]+df.name):
+                            print("PRINT ", os.getcwd(), "remove " + df.path[len(dpath):]+df.name)
                             os.remove(df.path[len(dpath):]+df.name)
                     else:
+                        print("PRINT ", os.getcwd(), "git checkout " + df.path[len(dpath):]+df.name)
                         os.system("git checkout " + df.path[len(dpath):]+df.name)
                 # Second, Checkout all other Modules/Plugins files. There are 2 verisoned files (.plan et .json build)
                 # so we remove them. For the not versioned files we use 'git checkout'
@@ -791,11 +792,12 @@ class RestCheckout(Resource):
                         else:
                             return  # module/plugin must have at least one file
                     for f in mfiles:
-                        print("PRINT ", os.getcwd(), "git checkout " + f.path[len(mpath):]+f.name)
-                        if f.is_versioned():
+                        if f.is_versioned() and "master.SNAPSHOT" not in df.name:
                             if os.path.isfile(f.path[len(mpath):]+f.name):
+                                print("PRINT ", os.getcwd(), "remove " + f.path[len(mpath):]+f.name)
                                 os.remove(f.path[len(mpath):]+f.name)
                         else:
+                            print("PRINT ", os.getcwd(), "git checkout " + f.path[len(mpath):]+f.name)
                             os.system("git checkout " + f.path[len(mpath):]+f.name)
                     if not isinstance(m, ariane_delivery.SubModule):
                         for s in m.list_submod:
